DocGenius-Revolutionizing-PDFs-with-AI
About the Project
This is a Python application that allows you to load a PDF and ask questions about it using natural language. The application uses a LLM to generate a response about your PDF. The LLM will not answer questions unrelated to the document. The application reads the PDF and splits the text into smaller chunks that can be then fed into a LLM. It uses OpenAI embeddings to create vector representations of the chunks. The application then finds the chunks that are semantically similar to the question that the user asked and feeds those chunks to the LLM to generate a response. 
Here is the Proof of Concept.
Required Libraries
tiktoken
faiss-cpu
langchain
PyPDF2
python-dotenv
streamlit